WIP move stuff around...
make it work with legacy stuff...

move stuff around and fix it

WIP move stuff around...

fix where stuff lives
